# interfaces/generated/types.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package generated

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Profile defines model for Profile.
type Profile struct {
	// Affiliation 所属
	Affiliation *string `json:"affiliation,omitempty"`

	// Bio 自己紹介
	Bio *string `json:"bio,omitempty"`

	// CreatedAt 作成日時
	CreatedAt *time.Time `json:"createdAt,omitempty"`

	// Id プロフィールのUUID
	Id *openapi_types.UUID `json:"id,omitempty"`

	// InstagramId InstagramのID
	InstagramId *string `json:"instagramId,omitempty"`

	// Name 名前
	Name *string `json:"name,omitempty"`

	// TwitterId TwitterのID
	TwitterId *string `json:"twitterId,omitempty"`
}

// ProfileInput defines model for ProfileInput.
type ProfileInput struct {
	// Affiliation 所属
	Affiliation *string `json:"affiliation,omitempty"`

	// Bio 自己紹介
	Bio *string `json:"bio,omitempty"`

	// InstagramId InstagramのID
	InstagramId *string `json:"instagramId,omitempty"`

	// Name 名前
	Name string `json:"name"`

	// TwitterId TwitterのID
	TwitterId *string `json:"twitterId,omitempty"`
}

// CreateProfileJSONRequestBody defines body for CreateProfile for application/json ContentType.
type CreateProfileJSONRequestBody = ProfileInput

// UpdateProfileJSONRequestBody defines body for UpdateProfile for application/json ContentType.
type UpdateProfileJSONRequestBody = ProfileInput